+++ tests/full_test_runner.py
@@ -15,6 +15,7 @@ def full_test(file: str,
               lost_count: int = 10000,
               lost_rule_count: int = 10000,
               error_on_space: bool = True,
+              error_on_no_rule: bool = True,
               br_builtins: bool = True,
               core_dump: bool = False,
               yields: bool = False,) -> backrooms.portal.Portal:
@@ -25,6 +26,7 @@ def full_test(file: str,
     :param lost_count: int
     :param lost_rule_count: int
     :param error_on_space: bool
+    :param error_on_no_rule: bool
     :param br_builtins: bool
     :param core_dump: bool
     :param yields: bool
@@ -41,6 +43,7 @@ def full_test(file: str,
                                            lost_count=lost_count,
                                            lost_rule_count=lost_rule_count,
                                            error_on_space=error_on_space,
+                                           error_on_no_rule=error_on_no_rule,
                                            br_builtins=br_builtins,
                                            core_dump=core_dump,
                                            yields=yields)